add Create Order Function

+++ models/order.js
@@ -7,7 +7,7 @@ module.exports = (sequelize, DataTypes) => {
       order_status: DataTypes.STRING,
       shipping_status: DataTypes.STRING,
       payment_status: DataTypes.STRING,
-      total_amount: DataTypes.FLOAT,
+      total_amount: DataTypes.INTEGER,
       name: DataTypes.STRING,
       address: DataTypes.STRING,
       email: DataTypes.STRING,
